src/requirements.txt
cython==0.29.1
pillow==5.3.0
pytesseract==0.2.4
tesserocr==2.3.1
boto3==1.9.71